requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
aniso8601==9.0.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
crcmod==1.7
cryptography==3.4.7
distlib==0.3.3
Django==3.2.7
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
django-oss-storage==1.1.1
django-sortedm2m==3.0.2
django-sortedone2many==0.2.2
djangorestframework==3.12.4
filelock==3.3.0
graphene==3.0b8
graphene-django==3.0.0b7
graphene-django-extras==0.5.2
graphene-django-pagination @ https://github.com/instruct-br/graphene-django-pagination/archive/master.zip
graphql-core==3.1.6
graphql-relay==3.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
jmespath==0.10.0
oss2==2.15.0
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.4.0
promise==2.3
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.8.1
virtualenv-clone==0.5.7
zipp==3.6.0
